requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
gunicorn==21.2.0
